// src/generated/model/getStudentByClassResponseDto.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * ITC Summer
 * ITCancún summer API
 * OpenAPI spec version: 1.0
 */
import type { StudentResponseDto } from './studentResponseDto';

export interface GetStudentByClassResponseDto {
  id_student_class: number;
  group_leader: boolean;
  student: StudentResponseDto;
}